Todo

Description

This module contains the TodoAPI and its implementation.

Documentation

type TodoAPI = "todo" :> (Get '[JSON] [Todo] :<|> (("addCategory" :> (QueryParam "category" String :> Post '[JSON] (Key Category))) :<|> (("add" :> (ReqBody '[JSON] Todo :> Get '[JSON] (Key Todo))) :<|> ((Capture "id" Int64 :> Get '[JSON] Todo) :<|> (("remove" :> (Capture "id" Int64 :> Get '[JSON] ())) :<|> ("check" :> Get '[JSON] [Todo]))))))

The Rest API hadling todos. All its services' uri- paths start with "todo"

todoServer

Arguments

:: ConnectionPool
the pool of database connections to be used

-> Server TodoAPI
the server for the TodoAPI

Implementation of the TodoAPI

checkTodoDeadline'

Arguments

:: [Todo]
List of Todos to check

-> Day
the day on wich the todos should be finished

-> [Todo]
list of todos with not enought time left

Checks for all given Todos whether or not there is enought time left for them to be compleated TODO: use duration

getToday

Arguments

:: IO Day
the Monad of today

Returns today as an IO Monad of Day