Types

Description

This module contains the custom database types which for some reason can't be in the same file as the database entities. Also there are the types to derive ToJSON and FromJSON which are only used for the API and not as database entities.

Documentation

data Status

Status of a Todo

Constructors

Open
not yet started

InProgress
started but not finished

Closed
finished

data Priority

Priority of a Todo

Constructors

Low
Low Priority

Middle
Medium Priority

High
High Priority

data Dependency

A Dependency element to be used as JSON for the TaskAPI

Constructors

Dependency

Fields

depends :: Int

the key of the task to depend on

major :: Int

major release number

minor :: Int

mionr release number

data Task

A Task element to be used as JSON for the TaskAPI

Constructors

Task

Fields

name :: String

the name of the task

dependencies :: [Dependency]